requirements.txt
PyPDF2
python-dotenv
requests
pandas
bs4
google-genai
supabase
httpx
fastapi
uvicorn
pydantic
scikit-learn